+++ assets/javascripts/kitten/components/text-copy/test.js
@@ -22,4 +22,18 @@ describe('<TextCopy />', () => {
       .toJSON()
     expect(component).toMatchSnapshot()
   })
+  it('should match snapshot with one line forced', () => {
+    const component = renderer
+      .create(
+        <TextCopy
+          alertMessage="Text copied!"
+          description="This is a description"
+          forceOneLine
+        >
+          Text To Copy
+        </TextCopy>,
+      )
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
 })